Move DummyModel to separate file
