+++ auto_causality/params.py
@@ -144,7 +144,7 @@ class SimpleParamService:
                 },
                 "defaults": {
                     "min_propensity": 1e-6,
-                    "n_estimators": 1000,
+                    "n_estimators": 1000,  # this is 1000 by default?
                     "min_samples_split": 5,
                     "min_samples_leaf": 5,
                     "min_weight_fraction_leaf": 0.0,
